Move common example code into mod
